feat(mcp): add chat button to registry beta catalog item detail page
