+++ platform/frontend/src/app/mcp/registry/beta/_parts/mcp-server-card.tsx
@@ -51,15 +51,10 @@ import {
 } from "@/components/ui/tooltip";
 import { TruncatedTooltip } from "@/components/ui/truncated-tooltip";
 import { LOCAL_MCP_DISABLED_MESSAGE } from "@/consts";
-import { fetchInternalAgents, useCreateProfile } from "@/lib/agent.query";
-import { useBulkAssignTools } from "@/lib/agent-tools.query";
 import { useHasPermissions, useSession } from "@/lib/auth/auth.query";
 import { useFeature } from "@/lib/config/config.query";
 import { useEnvironments } from "@/lib/environment.query";
-import {
-  fetchCatalogTools,
-  useReinstallInternalMcpCatalogItem,
-} from "@/lib/mcp/internal-mcp-catalog.query";
+import { useReinstallInternalMcpCatalogItem } from "@/lib/mcp/internal-mcp-catalog.query";
 import { useMcpServers } from "@/lib/mcp/mcp-server.query";
 import { useDefaultEnvironment } from "@/lib/organization.query";
 import { useAssignableTeams } from "@/lib/teams/team.query";
@@ -79,6 +74,7 @@ import {
 } from "../../_parts/uninstall-server-dialog";
 import { useCanReauthenticate } from "../../_parts/use-can-reauthenticate";
 import { CatalogEditNoAccess } from "./edit-catalog-dialog";
+import { useChatWithCatalogItem } from "./use-chat-with-catalog-item";
 
 export type CatalogItem =
   archestraApiTypes.GetInternalMcpCatalogResponses["200"][number];
@@ -138,9 +134,7 @@ export function McpServerCard({
 }: McpServerCardBaseProps) {
   const isPlaywrightVariant = isBuiltInPlaywright;
 
-  const createAgent = useCreateProfile();
-  const bulkAssignTools = useBulkAssignTools();
-  const [isChatCreating, setIsChatCreating] = useState(false);
+  const { startChat, isCreating: isChatCreating } = useChatWithCatalogItem();
 
   const isByosEnabled = useFeature("byosEnabled");
   const { data: session } = useSession();
@@ -258,50 +252,6 @@ export function McpServerCard({
       setEditNoAccessOpen(true);
     }
   }, [editParam, item.id, canEditCatalog, canEditCatalogLoading, router]);
-
-  const handleChatWithMcpServer = async () => {
-    setIsChatCreating(true);
-    const agentName = item.name;
-    try {
-      // Get or create: check if a personal agent with this name already exists for the current user
-      const existingAgents = await fetchInternalAgents();
-      const existing = existingAgents?.find(
-        (a) => a.name === agentName && a.authorId === currentUserId,
-      );
-
-      const agent =
-        existing ??
-        (await createAgent.mutateAsync({
-          name: agentName,
-          agentType: "agent",
-          scope: "personal",
-          teams: [],
-          icon: item.icon ?? undefined,
-        }));
-
-      const tools = await fetchCatalogTools(item.id);
-
-      if (agent && tools && tools.length > 0) {
-        const assignments = tools.map((tool) => ({
-          agentId: agent.id,
-          toolId: tool.id,
-          resolveAtCallTime: true,
-          ...(item.enterpriseManagedConfig
-            ? { credentialResolutionMode: "enterprise_managed" as const }
-            : {}),
-        }));
-        await bulkAssignTools.mutateAsync({ assignments });
-      }
-
-      if (agent) {
-        window.location.href = `/chat/new?agent_id=${agent.id}`;
-      }
-    } catch {
-      toast.error("Failed to create chat agent");
-    } finally {
-      setIsChatCreating(false);
-    }
-  };
 
   const mcpServerOfCurrentCatalogItem = allMcpServers?.filter(
     (s) => s.catalogId === item.id,
@@ -516,7 +466,7 @@ export function McpServerCard({
       size="sm"
       className="flex-1"
       disabled={isChatCreating}
-      onClick={handleChatWithMcpServer}
+      onClick={() => startChat(item)}
     >
       <MessageSquare className="h-4 w-4" />
       {isChatCreating ? "Creating..." : "Chat"}
